requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-pptx==0.6.22
PyMuPDF==1.23.8
pillow==9.5.0
pytesseract==0.3.10
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
